+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `-k / --kibibytes` option
 - Lookup `LS_BLOCK_SIZE` and `BLOCK_SIZE` environment variables to derive file block size.
 - `-b / --escape` option.
+- `--quoting-style=WORD` option.
 
 ### Changed